ContinuousMonitoring/grafana/client-component/pythonFlaskmicro/Helloapi/requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
